[Refactor] 타임스탬프 포매터 함수 분리
